refactor: remove old HeightmapGpu code
