+++ src/lib/index.ts
@@ -2,11 +2,7 @@ export { PromisesQueue } from './helpers/async/promises-queue';
 export { ELogLevel, setVerbosity } from './helpers/logger';
 export { HeightmapAtlas, type HeightmapAtlasStatistics } from './terrain/heightmap/atlas/heightmap-atlas';
 export { HeightmapViewerCpu, type HeightmapViewerCpuStatistics } from './terrain/heightmap/cpu/heightmap-viewer-cpu';
-export { HeightmapViewerGpu, type HeightmapViewerGpuStatistics } from './terrain/heightmap/gpu/new/heightmap-viewer-gpu';
-export {
-    HeightmapViewerGpu as HeightmapViewerGpuOld,
-    type HeightmapViewerGpuStatistics as HeightmapViewerGpuStatisticsOld,
-} from './terrain/heightmap/gpu/old/heightmap-viewer-gpu';
+export { HeightmapViewerGpu, type HeightmapViewerGpuStatistics } from './terrain/heightmap/gpu/heightmap-viewer-gpu';
 export type { HeightmapSamples, IHeightmap } from './terrain/heightmap/i-heightmap';
 export { type IHeightmapViewer } from './terrain/heightmap/i-heightmap-viewer';
 export { MaterialsStore } from './terrain/materials-store';
